+++ packages/jshint/index.js
@@ -1,9 +1,6 @@
-'use babel';
-
-/* globals atom */
-import { CompositeDisposable } from 'event-kit';
+/** @babel */
+import {CompositeDisposable} from 'event-kit';
 import path from 'path';
-import reactDomPragma from 'react-dom-pragma';
 import lazyRequire from 'lazy-req';
 
 const lazyReq = lazyRequire(require);
@@ -12,11 +9,14 @@ const jshint = lazyReq('jshint');
 const jsxhint = lazyReq('jshint-jsx');
 const cli = lazyReq('jshint/src/cli');
 const loadConfig = lazyReq('./load-config');
-const plugin = {};
 const markersByEditorId = {};
 const errorsByEditorId = {};
-const subscriptionTooltips = new CompositeDisposable();
+
+let subscriptionTooltips = new CompositeDisposable();
+let subscriptionEvents = new CompositeDisposable();
+
 let _;
+let statusBar;
 
 const SUPPORTED_GRAMMARS = [
 	'source.js',
@@ -24,17 +24,31 @@ const SUPPORTED_GRAMMARS = [
 	'source.js.jsx'
 ];
 
-const jsHintStatusBar = document.createElement('span');
+let currentLine;
+let currentChar;
+
+const goToError = () => {
+	const editor = atom.workspace.getActiveTextEditor();
+
+	if (!editor || !currentLine || !currentChar) {
+		return;
+	}
+
+	editor.setCursorBufferPosition([currentLine - 1, currentChar - 1]);
+};
+
+const jsHintStatusBar = document.createElement('a');
 jsHintStatusBar.setAttribute('id', 'jshint-statusbar');
 jsHintStatusBar.classList.add('inline-block');
+jsHintStatusBar.addEventListener('click', goToError);
 
 const updateStatusText = (line, character, reason) => {
 	jsHintStatusBar.textContent = line && character && reason ? `JSHint ${line}:${character} ${reason}` : '';
+	currentLine = line;
+	currentChar = character;
 };
 
-const getMarkersForEditor = () => {
-	const editor = atom.workspace.getActiveTextEditor();
-
+const getMarkersForEditor = editor => {
 	if (editor && markersByEditorId[editor.id]) {
 		return markersByEditorId[editor.id];
 	}
@@ -42,18 +56,15 @@ const getMarkersForEditor = () => {
 	return {};
 };
 
-const getErrorsForEditor = () => {
-	const editor = atom.workspace.getActiveTextEditor();
-
-	if (editor && errorsByEditorId[editor.id]) {
+const getErrorsForEditor = editor => {
+	if (errorsByEditorId[editor.id]) {
 		return errorsByEditorId[editor.id];
 	}
 
 	return [];
 };
 
-const destroyMarkerAtRow = row => {
-	const editor = atom.workspace.getActiveTextEditor();
+const destroyMarkerAtRow = (editor, row) => {
 	if (markersByEditorId[editor.id] && markersByEditorId[editor.id][row]) {
 		markersByEditorId[editor.id][row].destroy();
 		delete markersByEditorId[editor.id][row];
@@ -61,27 +72,29 @@ const destroyMarkerAtRow = row => {
 };
 
 const getRowForError = error => {
-	const line = error[0].line || 1; // JSHint reports `line: 0` when config error
+	// JSHint reports `line: 0` when config error
+	const line = error[0].line || 1;
+
 	const row = line - 1;
+
 	return row;
 };
 
-const clearOldMarkers = errors => {
+const clearOldMarkers = (editor, errors) => {
 	subscriptionTooltips.dispose();
+	subscriptionTooltips = new CompositeDisposable();
 
 	const rows = _.map(errors, error => getRowForError(error));
 
-	const oldMarkers = getMarkersForEditor();
+	const oldMarkers = getMarkersForEditor(editor);
 	_.each(_.keys(oldMarkers), row => {
 		if (!_.contains(rows, row)) {
-			destroyMarkerAtRow(row);
+			destroyMarkerAtRow(editor, row);
 		}
 	});
 };
 
-const saveMarker = (marker, row) => {
-	const editor = atom.workspace.getActiveTextEditor();
-
+const saveMarker = (editor, marker, row) => {
 	if (!markersByEditorId[editor.id]) {
 		markersByEditorId[editor.id] = {};
 	}
@@ -89,9 +102,7 @@ const saveMarker = (marker, row) => {
 	markersByEditorId[editor.id][row] = marker;
 };
 
-const getMarkerAtRow = row => {
-	const editor = atom.workspace.getActiveTextEditor();
-
+const getMarkerAtRow = (editor, row) => {
 	if (!markersByEditorId[editor.id]) {
 		return null;
 	}
@@ -100,7 +111,6 @@ const getMarkerAtRow = row => {
 };
 
 const updateStatusbar = () => {
-	const statusBar = atom.views.getView(atom.workspace).querySelector('.status-bar');
 	if (!statusBar) {
 		return;
 	}
@@ -110,6 +120,7 @@ const updateStatusbar = () => {
 	}
 
 	const editor = atom.workspace.getActiveTextEditor();
+
 	if (!editor || !errorsByEditorId[editor.id]) {
 		updateStatusText();
 		return;
@@ -117,50 +128,72 @@ const updateStatusbar = () => {
 
 	const line = editor.getCursorBufferPosition().row + 1;
 	let error = errorsByEditorId[editor.id][line] || _.first(_.compact(errorsByEditorId[editor.id]));
-	error = error[0];
+	error = Array.isArray(error) ? error[0] : {};
 
 	updateStatusText(error.line, error.character, error.reason);
 };
 
-const getReasonsForError = error => {
-	return _.map(error, el => `${el.character}: ${el.reason}`);
+const goToNextError = () => {
+	const editor = atom.workspace.getActiveTextEditor();
+
+	if (!editor || !markersByEditorId[editor.id] || !errorsByEditorId[editor.id]) {
+		return;
+	}
+
+	const cursorRow = editor.getCursorBufferPosition().row;
+
+	const markerRows = _.sortBy(_.map(_.keys(getMarkersForEditor(editor)), x => Number(x)));
+	let nextRow = _.find(markerRows, x => x > cursorRow);
+	if (!nextRow) {
+		nextRow = _.first(markerRows);
+	}
+	if (!nextRow) {
+		return;
+	}
+
+	const errors = errorsByEditorId[editor.id][nextRow + 1];
+	if (errors) {
+		editor.setCursorBufferPosition([nextRow, errors[0].character - 1]);
+	}
 };
 
+const getReasonsForError = error => {
+	return _.map(error, el => `${el.character}: ${el.reason} (${el.code})`);
+};
 
-const addReasons = (marker, error) => {
+const addReasons = (editor, marker, error) => {
 	const row = getRowForError(error);
-	const editorElement = atom.views.getView(atom.workspace.getActiveTextEditor());
+	const editorElement = atom.views.getView(editor);
 	const reasons = `<div class="jshint-errors">${getReasonsForError(error).join('<br>')}</div>`;
-
-	const target = editorElement.shadowRoot.querySelectorAll('.jshint-line-number.line-number-' + row);
+	const target = editorElement.shadowRoot.querySelectorAll(`.jshint-line-number.line-number-${row}`);
 	const tooltip = atom.tooltips.add(target, {
 		title: reasons,
 		placement: 'bottom',
 		delay: {show: 200}
 	});
+
 	subscriptionTooltips.add(tooltip);
 };
 
-const displayError = error => {
-	const row = getRowForError(error);
+const displayError = (editor, err) => {
+	const row = getRowForError(err);
 
-	if (getMarkerAtRow(row)) {
+	if (getMarkerAtRow(editor, row)) {
 		return;
 	}
 
-	const editor = atom.workspace.getActiveTextEditor();
 	const marker = editor.markBufferRange([[row, 0], [row, 1]]);
 	editor.decorateMarker(marker, {type: 'line', class: 'jshint-line'});
 	editor.decorateMarker(marker, {type: 'line-number', class: 'jshint-line-number'});
-	saveMarker(marker, row);
-	addReasons(marker, error);
+	saveMarker(editor, marker, row);
+	addReasons(editor, marker, err);
 };
 
-const displayErrors = () => {
-	const errors = _.compact(getErrorsForEditor());
-	clearOldMarkers(errors);
+const displayErrors = editor => {
+	const errors = _.compact(getErrorsForEditor(editor));
+	clearOldMarkers(editor, errors);
 	updateStatusbar();
-	_.each(errors, displayError);
+	_.each(errors, err => displayError(editor, err));
 };
 
 const removeMarkersForEditorId = id => {
@@ -195,23 +228,20 @@ const lint = () => {
 	// Remove errors and don't lint if file is ignored in .jshintignore
 	if (file && cli().gather({args: [file]}).length === 0) {
 		removeErrorsForEditorId(editor.id);
-		displayErrors();
+		displayErrors(editor);
 		removeMarkersForEditorId(editor.id);
 		return;
 	}
 
 	const config = file ? loadConfig()(file) : {};
-
 	const linter = (atom.config.get('jshint.supportLintingJsx') || atom.config.get('jshint.transformJsx')) ? jsxhint().JSXHINT : jshint().JSHINT;
 
-	const origCode = editor.getText();
-	const grammarScope = editor.getGrammar().scopeName;
-	const isJsx = grammarScope === 'source.jsx' || grammarScope === 'source.js.jsx';
-	const code = isJsx ? reactDomPragma(origCode) : origCode;
-	const pragmaWasAdded = code !== origCode;
+	if (Object.keys(config).length === 0 && atom.config.get('jshint.onlyConfig')) {
+		return;
+	}
 
 	try {
-		linter(code, config, config.globals);
+		linter(editor.getText(), config, config.globals);
 	} catch (err) {}
 
 	removeErrorsForEditorId(editor.id);
@@ -223,10 +253,6 @@ const lint = () => {
 		// aggregate same-line errors
 		const ret = [];
 		_.each(errors, el => {
-			if (pragmaWasAdded) {
-				el.line--;
-			}
-
 			const l = el.line;
 
 			if (Array.isArray(ret[l])) {
@@ -241,7 +267,7 @@ const lint = () => {
 		errorsByEditorId[editor.id] = ret;
 	}
 
-	displayErrors();
+	displayErrors(editor);
 };
 
 let debouncedLint = null;
@@ -249,39 +275,40 @@ let debouncedDisplayErrors = null;
 let debouncedUpdateStatusbar = null;
 
 const registerEvents = () => {
-	lint();
-	const workspaceElement = atom.views.getView(atom.workspace);
+	subscriptionEvents.dispose();
+	subscriptionEvents = new CompositeDisposable();
 
-	debouncedLint = debouncedLint || _.debounce(lint, 50);
-	debouncedDisplayErrors = debouncedDisplayErrors || _.debounce(displayErrors, 200);
-	debouncedUpdateStatusbar = debouncedUpdateStatusbar || _.debounce(updateStatusbar, 100);
+	updateStatusbar();
 
-	atom.workspace.observeTextEditors(editor => {
-		const buffer = editor.getBuffer();
+	const editor = atom.workspace.getActiveTextEditor();
+	if (!editor) {
+		return;
+	}
 
-		editor.emitter.off('scroll-top-changed', debouncedDisplayErrors);
-		editor.emitter.off('did-change-cursor-position', debouncedUpdateStatusbar);
-		buffer.emitter.off('did-save did-change-modified', debouncedLint);
+	displayErrors(editor);
 
-		if (!atom.config.get('jshint.validateOnlyOnSave')) {
-			buffer.onDidChangeModified(debouncedLint);
-		}
+	if (!atom.config.get('jshint.validateOnlyOnSave')) {
+		subscriptionEvents.add(editor.onDidChange(debouncedLint));
+		debouncedLint();
+	}
 
-		buffer.onDidSave(debouncedLint);
+	subscriptionEvents.add(editor.onDidSave(debouncedLint));
+	subscriptionEvents.add(editor.onDidChangeScrollTop(() => debouncedDisplayErrors(editor)));
+	subscriptionEvents.add(editor.onDidChangeCursorPosition(debouncedUpdateStatusbar));
 
-		editor.onDidChangeScrollTop(debouncedDisplayErrors);
-		editor.onDidChangeCursorPosition(debouncedUpdateStatusbar);
-	});
-
-	workspaceElement.addEventListener('editor:will-be-removed', (e, editorView) => {
-		if (editorView && editorView.editor) {
-			removeErrorsForEditorId(editorView.editor.id);
-			removeMarkersForEditorId(editorView.editor.id);
-		}
-	});
+	subscriptionEvents.add(editor.onDidDestroy(() => {
+		removeErrorsForEditorId(editor.id);
+		displayErrors(editor);
+		removeMarkersForEditorId(editor.id);
+	}));
 };
 
-export const config = plugin.config = {
+export const config = {
+	onlyConfig: {
+		type: 'boolean',
+		default: false,
+		description: 'Disable linter if there is no config file found for the linter.'
+	},
 	validateOnlyOnSave: {
 		type: 'boolean',
 		default: false
@@ -293,11 +320,26 @@ export const config = plugin.config = {
 	}
 };
 
-export const activate = plugin.activate = () => {
+let subscriptionMain = null;
+
+export const activate = () => {
 	_ = lodash();
-	registerEvents();
-	atom.config.observe('jshint.validateOnlyOnSave', registerEvents);
-	atom.commands.add('atom-workspace', 'jshint:lint', lint);
+	debouncedLint = _.debounce(lint, 200);
+	debouncedDisplayErrors = _.debounce(displayErrors, 200);
+	debouncedUpdateStatusbar = _.debounce(updateStatusbar, 100);
+
+	subscriptionMain = new CompositeDisposable();
+	subscriptionMain.add(atom.workspace.observeActivePaneItem(registerEvents));
+	subscriptionMain.add(atom.config.observe('jshint.validateOnlyOnSave', registerEvents));
+	subscriptionMain.add(atom.commands.add('atom-workspace', 'jshint:lint', lint));
+	subscriptionMain.add(atom.commands.add('atom-workspace', 'jshint:go-to-error', goToError));
+	subscriptionMain.add(atom.commands.add('atom-workspace', 'jshint:go-to-next-error', goToNextError));
 };
 
-export default plugin;
+export const deactivate = () => {
+	subscriptionTooltips.dispose();
+	subscriptionEvents.dispose();
+	subscriptionMain.dispose();
+};
+
+export const consumeStatusBar = instance => statusBar = instance;